third_party/blink/web_tests/wpt_internal/html/canvas/drawElementImage/draw-element-image-returned-matrix.html

<!DOCTYPE html>
<html>
<title>drawElementImage should return a matrix transforming the element's layout location to the drawn location</title>
<link rel="help" href="https://github.com/WICG/html-in-canvas">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<canvas id="canvas" style="width: 100px; height: 100px;" width="100" height="100" layoutsubtree>
  <div id="element" style="width: 10px; height: 10px;"></div>
</canvas>
<script>
  test(() => {
    let draw_transform = canvas.getContext('2d').drawElementImage(element, 0, 0);
    let expected = new DOMMatrix();
    assert_array_equals(draw_transform.toFloat64Array(), expected.toFloat64Array());
  }, 'Draw transform at the origin should be an identity matrix');

  test(() => {
    let draw_transform = canvas.getContext('2d').drawElementImage(element, 17, -9);
    let expected = new DOMMatrix().translateSelf(17, -9);
    assert_array_equals(draw_transform.toFloat64Array(), expected.toFloat64Array());
  }, 'Draw transform should account for x and y');

  test((t) => {
    t.add_cleanup(() => {
      element.style.transformOrigin = '';
    });
    element.style.transformOrigin = '0 0';
    let draw_transform = canvas.getContext('2d').drawElementImage(element, 0, 0, 20, 5);
    let expected = new DOMMatrix().scaleSelf(2, 0.5);
    assert_array_equals(draw_transform.toFloat64Array(), expected.toFloat64Array());
  }, 'Draw transform should account for width and height scale');

  test((t) => {
    t.add_cleanup(() => {
      element.style.transform = '';
    });
    element.style.transform = 'rotate(45deg)';
    let draw_transform = canvas.getContext('2d').drawElementImage(element, 0, 0);
    let expected = new DOMMatrix();
    assert_array_equals(draw_transform.toFloat64Array(), expected.toFloat64Array());
  }, 'Draw transform should not include transform on the element');

  test((t) => {
    t.add_cleanup(() => {
      canvas.getContext('2d').reset();
    });
    canvas.getContext('2d').translate(42, 17);
    let draw_transform = canvas.getContext('2d').drawElementImage(element, 0, 0);
    let expected = new DOMMatrix().translateSelf(42, 17);
    assert_array_equals(draw_transform.toFloat64Array(), expected.toFloat64Array());
  }, 'Draw transform should account for the current transform matrix with translation');

  test((t) => {
    t.add_cleanup(() => {
      canvas.style.zoom = '';
    });
    canvas.style.zoom = '1.5';
    let draw_transform = canvas.getContext('2d').drawElementImage(element, 10, 10);
    let expected = new DOMMatrix().translateSelf(10, 10);
    assert_array_equals(draw_transform.toFloat64Array(), expected.toFloat64Array());
  }, 'Draw transform should not be affected by zoom on the canvas');

  test((t) => {
    t.add_cleanup(() => {
      element.style.transformOrigin = '';
      canvas.getContext('2d').reset();
    });
    element.style.transformOrigin = '0 0';
    canvas.getContext('2d').rotate(Math.PI / 4);
    let draw_transform = canvas.getContext('2d').drawElementImage(element, 0, 0);
    let expected = new DOMMatrix().rotateSelf(45);
    assert_array_approx_equals(draw_transform.toFloat64Array(), expected.toFloat64Array(), 0.001);
  }, 'Draw transform should account for the current transform matrix with rotation');

  test((t) => {
    t.add_cleanup(() => {
      element.style.transformOrigin = '';
      canvas.getContext('2d').reset();
    });
    element.style.transformOrigin = '5px 5px';
    canvas.getContext('2d').rotate(Math.PI / 4);
    let draw_transform = canvas.getContext('2d').drawElementImage(element, 0, 0);
    let expected = new DOMMatrix()
      .translateSelf(-5, -5)
      .rotateSelf(45)
      .translateSelf(5, 5);
    assert_array_approx_equals(draw_transform.toFloat64Array(), expected.toFloat64Array(), 0.001);
  }, 'Draw transform should account for transform-origin');

  test((t) => {
    t.add_cleanup(() => {
      canvas.width = 100;
      canvas.height = 100;
    });
    canvas.width = 200;
    canvas.height = 50;
    let draw_transform = canvas.getContext('2d').drawElementImage(element, 0, 0);
    let expected = new DOMMatrix();
    assert_array_equals(draw_transform.toFloat64Array(), expected.toFloat64Array());
  }, 'Draw transform should account for canvas pixel scale');

  test((t) => {
    t.add_cleanup(() => {
      canvas.width = 100;
      canvas.height = 100;
      canvas.style.zoom = '';
      canvas.getContext('2d').reset();
      element.style.transformOrigin = '0 0';
    });
    canvas.width = 200;
    canvas.height = 50;
    canvas.style.zoom = 1.5;
    canvas.getContext('2d').translate(42, 17);
    canvas.getContext('2d').rotate(Math.PI / 4);
    element.style.transformOrigin = '5px 5px';
    let draw_transform = canvas.getContext('2d').drawElementImage(element, 13, -9);
    let expected = new DOMMatrix()
      .translateSelf(-5, -5)
      .scaleSelf(0.5, 2)
      .translateSelf(42, 17)
      .rotateSelf(45)
      .translateSelf(13, -9)
      .scaleSelf(2, 0.5)
      .translateSelf(5, 5);
    assert_array_approx_equals(draw_transform.toFloat64Array(), expected.toFloat64Array(), 0.001);
  }, 'Draw transform should account for canvas pixel scale, zoom, transform-origin, CTM, and x and y');
</script>
